+++ server/modules/hello.js
@@ -1,9 +1,9 @@
-const {sendAlert} = require("../helpers/messangers")
+const {sendMessage} = require("../helpers/messangers")
 
 const processHello = () => {
     const message = `Mina Monitor started successful! Have nice work!`
 
-    sendAlert("HELLO", message, false)
+    sendMessage("HELLO", message, false)
 }
 
 module.exports = {